WK1: fix User-Agent override
The previous code was caching the UA in the hash list used for
specific sites but if the url provided on command line is redirected
to another place the cache would then store the wrong url.

So now, simply store the UA in the WebPage and if it's not empty use
it when the UA is asked from the ChromeClient.

For WK2 a different code path is used and the feature works already as
expected.