Add a reference QWebPluginFactory and an example plugin to the browser.
The plugin factory is set to the QWebPage and a plugin is registered for
the mime type x-qt-plugin, so the plugin gets instantiated when an
object with the mime tyoe is found.

The example plugin uses OpenGL to draw a red triangle over a blue background
and paints it to the appropriate position of the page.

The file plugin.html contains an example page.