Added js communication to the plugin.
When the plugin gets added to the scene, it retrieves its QWebFrame that
has been stored by webkit as a property. Then it uses addToJavaScriptWindowObject
to make itself visible to the js environment in that frame.

Also added a signal and a slot to the plugin and updated the html example
to use them. The example uses a slot to set a number and prints the value
when a signal is received.